server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.110.0
fonttools==4.49.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
mysqlclient==2.2.4
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
sqlmodel==0.0.16
starlette==0.36.3
typing_extensions==4.9.0
tzdata==2024.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
